requirements.txt
Django==4.0.1
djangorestframework==3.13.1
django-cors-headers==3.11.0
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==20.1.0
whitenoise==5.3.0
python-dotenv==0.19.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
lyricsgenius==3.0.1
requests==2.27.1
django-filter==21.1
Unidecode==1.3.2